+++ alias_functions/glp.sh
@@ -7,4 +7,4 @@
 
 logCmd="git log --date=short --pretty=format:'%C(bold yellow)%h%C(reset) %C(white)%ad %C(bold blue)(%ar)%C(reset) %C(bold green)%an %C(reset)%C(cyan)%s%C(reset)'"
 
-eval $logCmd $*
+eval "$logCmd $*"