+++ alias_functions/gstw.sh
@@ -6,4 +6,4 @@
 # git stash show stash@{<stash_id>}
 
 stash_id="$1"
-git stash show stash@{$stash_id}
+git stash show "stash@{$stash_id}"