+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 > Any trouble, please visit the [troubleshooting page](https://github.com/diogocavilha/fancy-git/blob/master/TROUBLESHOOTING.md)
 
+## v7.1.6
+
+- Fix code stye according to ShellCheck rules.
+- Remove commands to enable/disable tag name.
+- Fix minor bugs.
+
 ## v7.1.5
 
 - Update screenshots on readme file.